Updated min dist to player for monster populating
The minimum distance is now determined by a floodfill based on which
cells are blocking projectiles instead.

Walls, closed wooden doors, etc are blocking the flood, and monsters
can spawn behind them if they are not reached by the flood (even if
their coordinates are very near the player).

Floor, deep water, chasms, grates, open wooden doors, closed gates,
etc are not blocking the flood, and monsters behind them are therefore
susceptible to being considered "too close to the player".