Bugfix for summoned monster leaders
When summoned monsters disappeared (as summoned monsters eventually do),
this did not consider if they were the leader of other monsters. One
situation where this caused a problem was when a monster summoned by a
summoning trap could themselves summon other monsters. Then it would set
itself as leader of the monsters it summoned, and then later disappear,
leaving the summoned monsters with an invalid 'leader' pointer.