Also print property start message when more applied
Previously, a start message was only printed printed if the same
property was not already applied, for example "I am wounded!" was only
for the first wound ("WOUNDED(1)"), but not if more of the same existing
property was applied ("WOUND(2)" etc).

Now the game gives correct feedback by printing messages for each extra
wound, or extra spirit sapping, etc.