Use weighted choice when spawning monster items
Previously, when spawning random items in monster inventories
(e.g. scrolls) a uniform distribution was used to select the item. This
made items which were supposed to be rare too common to drop from killed
monsters. Now a weighted choice is made instead, based on the spawn
chance item data.