Show if a door is stuck while adjacent
Stuck doors are drawn with a different color, and have "stuck" added to
their names if the player is adjacent to the door.

This avoids situations like the player igniting dynamite to throw
through a door, only to find out that it is stuck (which you previously
would have no way of knowing without attempting to open the door,
possibly alerting monsters).