Fix monster HP randomization bug
Fixed a bug which caused the lowest possible HP value for a monster to
be calculated wrong, when randomizing a monster's HP on spawning.

For example, a monster with 22 HP in the monster data would start with a
HP value in the range of 5-27, which has a much lower minimum value than
intended . Now the range for the same monster is 17-22.