Improve player monster awareness handling
Player awareness of a monster is no longer counted down if the player
can see the monster.

This prevents some glitchy behavior such as printing a message about a
Clear Ooze seeping through a door, and then never showing the monster
because the player loses awareness of it before their turn starts.