Make number queries more convenient for players
When the game asks the player for a number (for example when dropping a
stack of items or setting numerical options), entering any number clears
the default value. The default value is drawn with gray background until
the player has modified the value.

This allows player to for example drop 4 Throwing Knives from a stack of
12, without having to erase the default value first (fewer key presses).