Fix tile contour problem
After the recent rendering refactoring, tiles were always drawn with a
black contour. This ruined the graphics when drawing with black
foreground on a non-black background - for example the black exclamation
mark on gray background drawn when the player is aware of a monster.

Now tiles are only drawn with contours when both the foreground and
background is non-black.