Fix multi-projectile ranged attack crash bug
When multiple projectiles were fired (Tommy Gun), there was a problem
when one of the projectiles killed a monster so that another monster
spawned (Shapeshifter true form), which could lead to the game crashing,
or wrong projectile behavior.

The problem was that the projectile attack data updated their aim level
to something different from the original aim level, since a new monster
with a different height spawned in the cell aimed at. Now the original
aim level is stored when the projectile firing starts, and this is used
to force this aim level as they continue travelling on their path.