Controller support mode
Added a "Controller Support" input mode (in addition to Default and Vi).

From the manual:

There is also a Controller Support mode, which can be enabled in the
options menu. This mode is intended for cases such as binding a
controller stick to the numpad's eight directional keys (see the
movement layout above). In this mode, you must press the [q] key while
holding a direction in order to execute a move. Only pressing [q]
without a direction executes a "wait" command. While a direction is
held, a small symbol appears to indicate the active input. Note that
this confirmation is also required when choosing a direction for actions
such as kicking.