Fix sneak calculation
Sneak chance was calculated wrong when the chance for success was
rolled. Only the sneak chance of the hiding actor had any effect, but
search skill of the searching actor, distance, light and darkness was
also supposed to affect the chance.

The value presented when viewing the monster was correct, but the ACTUAL
chance when rolling for success was wrong.

This bug has existed since v21.0.0.