Add actor property upgrade functionality
It is now possible to configure that a type of actor property (status
effect) is an upgrade of another property, so that they cannot both be
active at the same time.

I.e. if it is specified that property B s an upgrade of A, then:
* If B is applied while A is active, A is first removed.
* If A is applied while B is active, A is not applied.

In any of the above cases, the duration of B will be updated however
(typically to the longest duration between A and B).

Currently, "Doomed" is specified to be an upgrade of "Cursed".